+++ pystorcli2/cmdRunner.py
@@ -8,18 +8,30 @@
 
 import os
 import shutil
-import subprocess
+from subprocess import Popen, PIPE
 from . import exc
+from typing import List, Tuple
+
+
+class StorcliRet():
+    def __init__(self, stdout: str, stderr: str, returncode: int):
+        self.stdout = stdout
+        self.stderr = stderr
+        self.returncode = returncode
 
 
 class CMDRunner():
     """This is a simple wrapper for subprocess.Popen()/subprocess.run(). The main idea is to inherit this class and create easy mockable tests.
     """
 
-    def run(self, args, **kwargs) -> subprocess.CompletedProcess:
+    def run(self, args, **kwargs) -> StorcliRet:
         """Runs a command and returns the output.
         """
-        return subprocess.run(args, **kwargs)
+        proc = Popen(args, stdout=PIPE, stderr=PIPE, **kwargs)
+
+        _stdout, _stderr = [i for i in proc.communicate()]
+
+        return StorcliRet(_stdout, _stderr, proc.returncode)
 
     def binaryCheck(self, binary) -> str:
         """Verify and return full binary path