+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Version 0.6.4
 - [**Drive**]             Added some aliases to `drive.state` & `drive.set_state` method. (Issue [#12](https://github.com/Naudit/pystorcli2/issues/12))
 - Added enum-class StorcliError. This Enum stores any posible (documented) error reported by storcli. We can expect & catch some of them if we want to
 
+- **Changes**:
+  ----------
+  - [**Storcli**]         Migrated from subprocess.run -> subprocess.popen. Requested by @ulmitov in Issue [#13](https://github.com/Naudit/pystorcli2/issues/13)
+
 Thanks to @ulmitov & @dgilbert101 for the contributions to this release
 
 Version 0.6.3